+++ src/helpers/UppyFileTypeCheckerPlugin.js
@@ -1,13 +1,37 @@
-// import BasePlugin from '@uppy/core/lib/BasePlugin.js';
 import Uppy from '@uppy/core';
+
+import { audioToText, image_classification, image_enhancement, object_detection, semantic_segmentation } from './TaskIDs';
 
 import fileTypeChecker from "file-type-checker";
 
+export const getAllowedFileTypes = (task) => {
+    switch (task) {
+      case audioToText:
+        return {
+          fileTypes: ['aac', 'amr', 'flac', 'mp3', 'mp4', 'm4a', 'wav', 'webm'],
+          mimeTypes: ['audio/*', 'video/*'],
+        };
+      case image_classification:
+      case image_enhancement:
+      case object_detection:
+      case semantic_segmentation:
+        return {
+          fileTypes: ['bmp', 'gif', 'ico', 'jpeg', 'pdf', 'png', 'psd'],
+          mimeTypes: ['image/*']
+        };
+      default:
+        // Allow all file types? Or disallow all file types?
+        return {
+          fileTypes: ['*'],
+          mimeTypes: '*/*'
+        };
+    }
+}
 
 export default class UppyFileTypeCheckerPlugin extends Uppy.Plugin {
 	constructor(uppy, opts) {
 		super(uppy, opts);
-        console.log('opts', opts)
+
 		this.id = opts.id || 'UppyFileTypeCheckerPlugin';
         // A type can be anything—some plugins use types to decide whether to do something to some other plugin.
 		// https://uppy.io/docs/guides/building-plugins/
@@ -15,24 +39,18 @@ export default class UppyFileTypeCheckerPlugin extends Uppy.Plugin {
 
         this.allowedFileTypes = opts.allowedFileTypes;
 
-        this.prepareUpload = this.prepareUpload.bind(this); // ← this!
-
         this.confirmFileType = this.confirmFileType.bind(this);
 	}
 
     confirmFileType = async (fileIDs) => {
+        // Note: This will break if we ever allow multiple uploads
         const file = this.uppy.getFile(fileIDs[0]);
-        // console.log('file', file)
-        const blob = new Blob([file.data])
-        // console.log('blob', blob)
-        const bufferResult = await blob.arrayBuffer()
-        // console.log(bufferResult)
-        const fileCheckerResult = fileTypeChecker.detectFile(bufferResult)
-        // console.log(fileCheckerResult)
-        console.log('allowed file types', this.allowedFileTypes)
+        const blob = new Blob([file.data]);
+        const bufferResult = await blob.arrayBuffer();
+        // Note: Doesn't seem like we need this additional check?
+        // const fileCheckerResult = fileTypeChecker.detectFile(bufferResult)
 
         const validation = fileTypeChecker.validateFileType(bufferResult, this.allowedFileTypes);
-        console.log('validation: ', validation)  
 
         if (validation) {
             return Promise.resolve()
@@ -41,32 +59,17 @@ export default class UppyFileTypeCheckerPlugin extends Uppy.Plugin {
             // Show error message to user
             this.uppy.info(`Something went wrong while adding "${file.data.name}". ` + 
             `Please check that the file extension is correct, or try a different file.`, 'error', 10000);  
-
+            // Print error in the console
             return Promise.reject(new Error(`"${file.data.name}" failed validation by file-type-checker. ` + 
             `Confirm the extension is correct and that it is in the list of allowed file types.`));
-        }
-
-
-
-        // return Promise.resolve();        
+        }      
     }    
 
-    async prepareUpload(fileIDs) {
-        console.log(fileIDs)
-		console.log(this); // `this` refers to the `MyPlugin` instance.
-
-
-		return Promise.resolve();
-	}
 	install() {
-		this.uppy.addPreProcessor(this.prepareUpload);
-
         this.uppy.addPreProcessor(this.confirmFileType);
 	}
 
 	uninstall() {
-		this.uppy.removePreProcessor(this.prepareUpload);
-
         this.uppy.removePreProcessor(this.confirmFileType);
 	}    
 }
